# codr-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Users demo
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
paths:
...
  /modules/{module_id}:
    x-swagger-router-controller: modules
    get:
      description: Returns the details of the specified module.
      operationId: getModuleById
      parameters:
        - name: module_id
          in: path
          description: ID of the module to fetch.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ModuleResponse"
        "404":
          description: Error, module could not be found.
          
    delete:
      description: Deletes the specified module.
      operationId: deleteModuleById
      parameters:
        - name: module_id
          in: path
          description: ID of the module to delete.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success.
        "401":
          description: Not allowed.
        "404":
          description: Error, module could not be found.
        "500":
          description: Internal server error
    put:
      description: Updates the specified module.
      operationId: putModuleById
      parameters:
        - in: body
          name: body
          description: Data for the module to update.
          schema:
            $ref: "#/definitions/ModuleUpdate"
        - name: module_id
          in: path
          description: ID of the module to update.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ModuleResponse"
        "401":
          description: Error, user is not allowed to access this page.
        "404":
          description: Error, module could not be found.
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
...
  /modules/{module_id}/courses:
    x-swagger-router-controller: courses
    get:
      description: Returns the courses of the specified module.
      operationId: getCoursesByModuleId
      parameters:
        - name: module_id
          in: path
          description: ID of the module from which to fetch courses.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
                  type: object
                  required:
                    - name
                    - description
                  properties:
                    id:
                      type: integer
                      format: int64
                    module_id:
                      type: integer
                      format: int64
                    name:
                      type: string
                    description:
                      type: string
                    can_edit:
                      type: boolean
        "401":
          description: Error, user is not allowed to access this page.
        "404":
          description: Error, module could not be found.
        "500":
          description: Internal server error
    post:
      description: Creates a new course for the given module.
      operationId: postCourseByModuleId
      parameters:
        - name: module_id
          in: path
          description: ID of the module from which to fetch courses.
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: JSON representation of the course to create.
          required: true
          schema:
            $ref: "#/definitions/CourseUpdate"
      responses:
        "201":
          description: Success, new course has been created.
          schema:
            $ref: "#/definitions/CourseResponse"
        "401":
          description: Error, user is not allowed to access this page.
        "405":
          description: Error, invalid input
        "500":
          description: Internal server error
  /modules/{module_id}/courses/{course_id}:
    x-swagger-router-controller: courses
    get:
      description: Returns the specified course for the specified module.
      operationId: getCourseById
      parameters:
        - name: module_id
          in: path
          description: ID of the module to fetch.
          required: true
          type: integer
          format: int64
        - name: course_id
          in: path
          description: ID of the course to fetch.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CourseResponseGet"
        "401":
          description: Error, user is not allowed to access this page.
        "404":
          description: Error, module or course could not be found.
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
    delete:
      description: Deletes the specified course from the specified module.
      operationId: deleteCourseById
      parameters:
        - name: module_id
          in: path
          description: ID of the module from which to delete a course.
          required: true
          type: integer
          format: int64
        - name: course_id
          in: path
          description: ID of the course to delete.
          required: true
          type: integer
          format: int64          
      responses:
        "200":
          description: Success.
        "401":
          description: Error, user is not allowed to access this page.
        "404":
          description: Error, module or course could not be found.
        "500":
          description: Internal server error
    put:
      description: Updates the specified course for the specified module.
      operationId: putCourseById
      parameters:
        - name: body
          in: body
          description: Data for the course to update.
          required: true
          schema:
            $ref: "#/definitions/CourseUpdate"
        - name: module_id
          in: path
          description: ID of the module from which to update a course.
          required: true
          type: integer
          format: int64
        - name: course_id
          in: path
          description: ID of the course to fetch.
          required: true
          type: integer
          format: int64  
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CourseResponse"
        "401":
          description: Error, user is not allowed to access this page.
        "404":
          description: Error, module or course could not be found.
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      display_name:
        type: string
...
  CommentResponseHead:
    type: object
    properties:
      id:
        type: integer
        format: int64
      parent_id:
        type: integer
        format: int64
      file_id:
        type: integer
        format: int64
      project_id:
        type: integer
        format: int64
      exercise_id:
        type: integer
        format: int64
      course_id:
        type: integer
        format: int64
      module_id:
        type: integer
        format: int64
      line_range:
        type: string
      content:
        type: string
      author_id:
        type: integer
        format: int64
      author_name:
        type: string
      visible:
        type: boolean
      timestamp:
        type: integer
        format: int64
      deleted:
        type: boolean
      can_edit:
        type: boolean
      can_delete:
        type: boolean
      can_toggle_visibility:
        type: boolean
      children:
        type: array
        items:
          $ref: '#/definitions/CommentResponseTail'
  CommentResponseTail:
    type: object
    properties:
      id:
        type: integer
        format: int64
      parent_id:
        type: integer
        format: int64
      file_id:
        type: integer
        format: int64
      project_id:
        type: integer
        format: int64
      exercise_id:
        type: integer
        format: int64
      course_id:
        type: integer
        format: int64
      module_id:
        type: integer
        format: int64
      line_range:
        type: string
      content:
        type: string
      author_id:
        type: integer
        format: int64
      author_name:
        type: string
      visible:
        type: boolean
      timestamp:
        type: integer
        format: int64
      deleted:
        type: boolean
      can_edit:
        type: boolean
      can_delete:
        type: boolean
      can_toggle_visibility:
        type: boolean
      children:
        type: array
        items:
          type: object
...
  Group:
    type: object
    properties:
      id:
        type: integer
        format: int64
      module_id:
        type: integer
        format: int64
      Users:
        type: array
        items: 
          $ref: '#/definitions/User'
...
  ModuleUpdate:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      users:
        type: array
        items:
          type: object
          properties:
            user_id:
              type: integer
              format: int64
            group_id:
              type: integer
              format: int64
...
  ModuleResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      can_edit:
        type: boolean
      courses:
        type: array
        items:
          $ref: '#/definitions/ModuleCourse'
      groups:
        type: array
        items:
          $ref: '#/definitions/Group'
...
  CourseUpdate:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
  CourseResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
  ModuleCourse:
    type: object
    required:
      - name
      - description
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string  
  CourseResponseGet:
    type: object
    required:
      - name
      - description
      - can_edit
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      can_edit:
        type: boolean
      exercises:
        type: array
        items:
          $ref: '#/definitions/Exercise'
  Exercise:
    type: object
    required:
      - rights_template_id
      - name
      - description
    properties:
      id:
        type: integer
        format: int64
      rights_template_id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      projects:
        type: array
        items:
          $ref: '#/definitions/Project'
  Project:
    type: object
    required:
      - projectgroup_id
    properties:
      id:
        type: integer
        format: int64
      projectgroup_id:
        type: integer
        format: int64
      files:
        type: array
        items:
          $ref: '#/definitions/File'
...
  File:
    type: object
    required:
      - id
      - path
      - content
      - comments
    properties:
      id:
        type: integer
        format: int64
      path:
        type: string
      content:
        type: array
        items:
          $ref: '#/definitions/Line'
      comments:
        type: array
        items:
          $ref: '#/definitions/CommentResponseHead'
...
  Line:
    type: string
  ErrorResponse:
    type: object
    properties:
      message:
        type: string